mark shader uniforms as changed on shader ini (bugfix for hot-reload)

+++ Renderer/src/main/java/darwin/resourcehandling/resmanagment/ShaderLoader.java
@@ -139,8 +139,10 @@ public class ShaderLoader {
                         ShaderProgramm compiledShader = util.compileShader(file);
                         ShaderProgramm old = shader.getProgramm();
                         shader.ini(compiledShader);
-                        old.delete();
                         logger.info("Shader " + name + " was succesfully compiled!");
+                        if (old != null) {
+                            old.delete();
+                        }
                     } catch (Throwable ex) {
                         logger.warn(ex.getLocalizedMessage());
                     }
